HAI-194 Refactor HaittojenhallintasuunnitelmaInfo into its own component and make styling uniform with application form summary page styling
